refactors duplicated functions to replayFunction.js
